# fix DynamicPPL.jl/pull/261