https://github.com/TuringLang/Libtask.jl/pull/new/DynamicPPL/pull/261

The issue was caused by the usage of WeakRef to an immutable object:

An immutable object could be moved under some circumstances, but the weak ref to it won't know it's moved, it will see moving as a GC, and mark its referencing value as nothing.

I changed the struct `TArray` from immutable to mutable to avoid this false elimination, and `TArray` only has two pointer fields, so if it is mutable or immutable doesn't make much difference, I think.