@yebai 2021-06-13
> I changed the struct TArray from immutable to mutable to avoid this false elimination, and TArray only has two pointer fields, so if it is mutable or immutable doesn't make much difference, I think.

It'll be helpful to have some basic benchmarking for the change, I think.