@KDr2 2021-06-14
> Can we add a test for the bug? I'll also open a separate PR with some integration tests for Turing and AdvancedPS but they don't help if Libtask is upper bounded explicitly.

I am not sure whether there's a certain way to cause an immutable object to be moved, thus we can construct a test case for this.  As how it is designed, the moving of objects is opaque to the Julia code and can't be explicitly controlled by Julia code, but I will try to construct one.