@KDr2 2021-06-14
> It'll be helpful to have some basic benchmarking for the change, I think.

I ran the old benchmarks, here are the results:
Before:
```
= Benchmarks on Arrays =
indexing:   2.554 ns (0 allocations: 0 bytes)
set indexing:   4.416 ns (0 allocations: 0 bytes)
broadcast:   20.374 ms (12 allocations: 45.78 MiB)
= Benchmarks on TArrays =
indexing:   137.360 ns (0 allocations: 0 bytes)
set indexing:   972.640 ns (0 allocations: 0 bytes)
broadcast:   21.410 ms (78 allocations: 45.78 MiB)
```
After:
```
= Benchmarks on Arrays =
indexing:   2.556 ns (0 allocations: 0 bytes)
set indexing:   4.708 ns (0 allocations: 0 bytes)
broadcast:   21.192 ms (12 allocations: 45.78 MiB)
= Benchmarks on TArrays =
indexing:   134.313 ns (0 allocations: 0 bytes)
set indexing:   813.130 ns (0 allocations: 0 bytes)
broadcast:   21.429 ms (60 allocations: 45.78 MiB)
```

But these are mostly about data accessing, rather than the placement and moving of `TArray` objects. We can finely control the moving of immutable objects so it's hard to construct a benchmark about that. 

I will do a benchmark if I can successfully construct a test case for this issue. 

After all, I guess the move may occur when a ctask is copied, but not every time a ctask is copied. So its usage in this aspect is not frequent. Meantime, the underlying array data is not directly in the `TArray` so the placement of it doesn't change during this commit. So I think the performance affection is really trivial, either from the perspective of data locality or memory copying.