@KDr2 2021-06-15
I found that if you copy a task a big amount of times the immutable objects in the task will be moved, and wrote a test case according to it.

It seems not impossible to benchmark it because the old version will throw an error if you run a benchmark against it.