@yebai 2021-06-15
Thanks, @KDr2 - ready to merge this PR. Could you add the benchmark scripts to the repo (e.g. `benchmarks/scripts.jl`) as a follow-up PR?